refactor(ceis_dashboard): separation of concerns
