+++ clab_ceis/ceis_dashboard/callbacks/api.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import requests
 
-from clab_ceis import config
-from clab_ceis.ceis_backend.models import Co2Response
+import config
+from ceis_backend.models import Co2Response
+# from clab_ceis.ceis_backend.models import Co2Response
 
 
 def fetch_fabric_blocks():
