+++ clab_ceis/ceis_dashboard/ceis_callbacks.py
@@ -1,7 +1,7 @@
 from dash import Dash
 
-from clab_ceis.ceis_dashboard import ceis_data
-from clab_ceis.ceis_dashboard.callbacks import (
+import ceis_data
+from callbacks import (
     register_co2_callbacks,
     register_dashboard_table_callbacks,
     register_fabric_block_callbacks,
